LPS-203925 Merge portal-impl LayoutSetModelListener into layout-impl LayoutSetModelListener
